feat: Board 인터페이스 정의
