+++ src/main/java/kr/codesqaud/chessgame/chess/Rank.java
@@ -4,17 +4,13 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
+import kr.codesqaud.chessgame.pieces.Color;
 import kr.codesqaud.chessgame.pieces.Piece;
-import kr.codesqaud.chessgame.pieces.Piece.Color;
-import kr.codesqaud.chessgame.pieces.Piece.Type;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import kr.codesqaud.chessgame.pieces.Type;
 
 public class Rank implements Comparable<Rank> {
 
-    private static final Logger logger = LoggerFactory.getLogger(Rank.class);
-
-    private final List<Piece> pieces = new ArrayList<>(Board.SIZE);
+    private final List<Piece> pieces = new ArrayList<>(ChessBoard.SIZE);
     private final int value;
 
     public Rank(final int value) {
